Add `make run` to launch the enclave
Record the command for launching the enclave in debug mode with
log streaming. There's no corresponding `make stop` because
closing the client ssh connection terminates the application
anyway.

Note that to deploy with `--memory 4000` the allocation limit
must first be raised in `/etc/nitro_enclaves/allocator.yaml`
on the parent instance.